Split Dashboard components into several files; add load more functionality.
