+++ src/components/modules/all.js
@@ -1,4 +1,4 @@
-import Dashboard from 'src/components/modules/Dashboard';
+import Dashboard from 'src/components/modules/dashboard/ApolloDashboard';
 import Database from 'src/components/modules/Database';
 import Doujin from 'src/components/modules/Doujin';
 import Forum from 'src/components/modules/Forum';
